requirements.txt
line-bot-sdk==3.7
flask==3.0.0
azure-ai-translation-text==1.0.0